refactor: add pkg/k8s
